+++ CHANGELOG.md
@@ -22,6 +22,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - The internal `.wireit/*/state` file was renamed to `.wireit/*/fingerprint`.
   Should have no effect.
 
+- If a script does not define a `"command"`, then fingerprints, lock files, and
+  cache entries are no longer written to the `.wireit` directory. This change
+  should have no user-facing effect apart from a very minor performance
+  improvement.
+
 ## [0.4.3] - 2022-05-15
 
 ### Changed